requirements.txt
mysqlclient
psycopg2
python-dotenv